backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
